% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze}
\usage{
analyze(
  expression_data,
  regressor_matrix,
  ref,
  type_1_cols,
  type_2_cols,
  tfbm_ref,
  exp_symbol_col = 1,
  reg_id_col = 1,
  ref_symbol_col = 1,
  foldThreshDEG = 1.5,
  DEGfun = NULL,
  screenSD = 0,
  logZeroSub = 0.001,
  n_boot = 200,
  verbose = TRUE
)
}
\arguments{
\item{expression_data}{a data frame of gene expression values in long format, where one column is gene
symbols (the first column by default) and other all other columns are subject IDs.
See \code{data("TAU_Trials3_Gene_CPM_Log2")} for an example of appropriate formatting.}

\item{regressor_matrix}{a matrix containing subject IDs, a predictor of interest, and covariates of interest (if any).
The subject ID column position may be altered using the \code{reg_id_col} parameter, but the predictor of interest
column must precede that of covariates in the matrix. See \code{data("TAUTrials2022BC_Intervention_Rm1BadQC_RmB14")}
for an example of appropriate formatting.}

\item{ref}{a data frame in long format containing gene expression value of the reference data frame.
See \code{data("HumanM1M2_3Reps_Martinez")} for an example of appropriate formatting.}

\item{type_1_cols}{an integer vector indicating the position of replicates for the first "cell type"/"subject type".}

\item{type_2_cols}{an integer vector indicating the position of replicates for the second "cell type"/"subject type".}

\item{tfbm_ref}{a list object containing several tfbm reference data frames;
the first column of each data frame must contain gene symbols and be named "gene"}

\item{exp_symbol_col}{an integer indicating the position of the gene symbol column in the expression data frame;
the first column is used by default.}

\item{reg_id_col}{an integer indicating the position of of the subject ID column in the \code{regressor_matrix}}

\item{ref_symbol_col}{an integer indicating the position of the gene symbol column in the reference data frame;
the first column is used by default.}

\item{foldThreshDEG}{the X-fold expression threshold to be exceeded for a given gene to be
considered differentially expressed.}

\item{DEGfun}{optionally input a custom function to derived differentially expressed genes. This function pulls variables
from the local environment (including df_DEG, foldThreshDEG) and MUST return a DEG vector; see examples for more.}

\item{screenSD}{minimum gene expression SD; genes that show expression variability below
threshold are excluded.}

\item{logZeroSub}{a numeric value used to substitute expression values in the reference data frame that cannot be
log transformed; .001 is used by default.}

\item{n_boot}{the number of bootstrap sample to run.}

\item{verbose}{a bool indicating whether to print function outputs/info to the console.}
}
\value{
toa returns a list object containing:
\enumerate{
\item a DEG result object.
\item a toa result object.
\item a toa_boot result object.
\item a tfbm result object.
\item a tfbm_boot result object.
}
}
\description{
performs primary analyses from the toa package in succession so as to identify differentially expression genes, then perform
transcript origin analysis (or toa), then perform the transcription factor binding motif variant of toa.
}
\examples{
\dontrun{
#
library(Rfssa)

#load example data
data("TAU_Trials3_Gene_CPM_Log2")
data("TAUTrials2022BC_Intervention_Rm1BadQC_RmB14")
data("HumanM1M2_3Reps_Martinez")
load_github_data("https://github.com/amanigaultw/TELiS/blob/main/HumanTransfacTELiS2019.RData")

#analyze
analyze_result <- analyze(expression_data = TAU_Trials3_Gene_CPM_Log2,
                          regressor_matrix = TAUTrials2022BC_Intervention_Rm1BadQC_RmB14,
                          ref = HumanM1M2_3Reps_Martinez,
                          type_1_cols = 2:4,
                          type_2_cols = 5:7,
                          tfbm_ref = HumanTransfacTELiS2019,
                          foldThreshDEG = 2)
}
}
