% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CGCD14_Group}
\alias{CGCD14_Group}
\title{CGCD14_Group dataset}
\format{
An object of class \code{data.frame} with 80 rows and 2 columns.
}
\usage{
CGCD14_Group
}
\description{
CGCD14_Group dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{internal}
