% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CGCD14_RNADataLog2}
\alias{CGCD14_RNADataLog2}
\title{CGCD14_RNADataLog2 dataset}
\format{
An object of class \code{data.frame} with 34551 rows and 81 columns.
}
\usage{
CGCD14_RNADataLog2
}
\description{
CGCD14_RNADataLog2 dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{internal}
