% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Chang}
\alias{Chang}
\title{Chang dataset}
\format{
An object of class \code{data.frame} with 6 rows and 18119 columns.
}
\usage{
Chang
}
\description{
Chang dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{internal}
