% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_boot.R
\name{contrast_boot}
\alias{contrast_boot}
\title{bootstrap test of gene contrast}
\usage{
contrast_boot(
  toa_boot_result = NULL,
  DEG_result = NULL,
  n_boot = NULL,
  contrast,
  verbose = TRUE
)
}
\arguments{
\item{toa_boot_result}{a toa_boot result object produced using \code{toa_boot()}.}

\item{DEG_result}{a DEG result object produced using \code{get_DEG()}.}

\item{n_boot}{the number of bootstrap sample to run.}

\item{contrast}{a dataframe containing gene symbols (1st column) and contrast loadings (2nd column).}

\item{verbose}{a bool indicating whether a progress bar should be shown.}
}
\value{
a list object containing:
\enumerate{
\item a results data frame.
\item a list of input arguments.
}
}
\description{
Computes a mean gene contrast score along with bootstrapped tests statistics. A single \code{toa_boot_result} object or the combination of \code{DEG_result} object and \code{n_boot} are acceptable inputs.
}
\examples{
\dontrun{
#load example data
data("TAU_Trials3_Gene_CPM_Log2")
data("TAUTrials2022BC_Intervention_Rm1BadQC_RmB14")
data("HumanM1M2_3Reps_Martinez")
data("CTRAProinflamAntiviralAntibody")

#get DEG
DEG_result <- get_DEG(expression_data = TAU_Trials3_Gene_CPM_Log2,
                      regressor_matrix = TAUTrials2022BC_Intervention_Rm1BadQC_RmB14,
                      foldThreshDEG = 2)

#get toa
toa_result <- toa(DEG_result = DEG_result,
                  ref = HumanM1M2_3Reps_Martinez,
                  type_1_cols = 2:4,
                  type_2_cols = 5:7)

#get bootstrapped stats
toa_boot_result <- toa_boot(toa_result)

#METHOD 1 (leverages existing bootstrap resamples)
contrast_boot_result_1 <- contrast_boot(toa_boot_result = toa_boot_result,
                                        contrast = CTRAProinflamAntiviralAntibody)

#METHOD 2 (generates new bootstrap resamples)
contrast_boot_result_2 <- contrast_boot(DEG_result = DEG_result,
                                        n_boot = 200,
                                        contrast = CTRAProinflamAntiviralAntibody)

}
}
