% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CTRAAntiviralAntibody}
\alias{CTRAAntiviralAntibody}
\title{CTRAAntiviralAntibody dataset}
\format{
An object of class \code{data.frame} with 34 rows and 2 columns.
}
\usage{
CTRAAntiviralAntibody
}
\description{
CTRAAntiviralAntibody dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{internal}
