% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CTRAProInflam}
\alias{CTRAProInflam}
\title{CTRAProInflam dataset}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\usage{
CTRAProInflam
}
\description{
CTRAProInflam dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{internal}
