% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CTRAProinflamAntiviralAntibody}
\alias{CTRAProinflamAntiviralAntibody}
\title{CTRAProinflamAntiviralAntibody dataset}
\format{
An object of class \code{data.frame} with 57 rows and 2 columns.
}
\usage{
CTRAProinflamAntiviralAntibody
}
\description{
CTRAProinflamAntiviralAntibody dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{internal}
