% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decouple_genes.R
\name{decouple_genes}
\alias{decouple_genes}
\title{decouples multi-gene rows}
\usage{
decouple_genes(df, sep = "///", gene_col = 1, header = TRUE, progress = TRUE)
}
\arguments{
\item{df}{a data frame or matrix containing gene symbols and gene expression values.}

\item{sep}{a string indicating how multi-gene symbols are separated (e.g., "///").}

\item{gene_col}{a numeric value indicating the position of the gene symbol column within \code{df}; the first column is used by default.}

\item{header}{bool indicating whether the first row of data is a header; TRUE by default.}

\item{progress}{bool indicating whether to display function progress.}
}
\value{
a data frame containing one row per gene symbols, where expression values are identical for previously coupled symbols.
}
\description{
produces a data frame containing one row per gene symbols, where expression values
are identical for previously coupled symbols.
}
\examples{
\dontrun{
#load example data
data("HumanCD14CD16NegVsPosVsDC_3Reps_Ziegler")
#decouple_genes
decoupled_data <- decouple_genes(HumanCD14CD16NegVsPosVsDC_3Reps_Ziegler)
}
}
