% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epith_mesen_ref_raw}
\alias{epith_mesen_ref_raw}
\title{epith_mesen_ref_raw dataset}
\format{
An object of class \code{data.frame} with 25684 rows and 21 columns.
}
\usage{
epith_mesen_ref_raw
}
\description{
epith_mesen_ref_raw dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{internal}
