% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_toa_ref.R
\name{get_toa_ref}
\alias{get_toa_ref}
\title{get reference data needed for Transcript Origin Analysis}
\usage{
get_toa_ref(
  gene_symbols,
  exp_treatment,
  exp_control,
  logZeroSub = 0.001,
  toupper = TRUE
)
}
\arguments{
\item{gene_symbols}{a character vector containing gene symbols.}

\item{exp_treatment}{a numeric matrix of gene expression values, where columns are observations from a treatment/exposure group.}

\item{exp_control}{a numeric matrix of gene expression values, where columns are observations from a control/reference group.}

\item{logZeroSub}{the numeric value used to replace expression values that cannot be log2 transformed; .001 by default.}

\item{toupper}{bool indicating whether gene symbols should be set to uppercase.}
}
\value{
a data frame containing unique gene symbols and average diagnosticity scores.
}
\description{
\code{get_toa_ref} computes diagnosticity scores for a given set of treatment v. control observations, typically replicates.
}
\examples{
#load example data
data("epith_mesen_ref_raw")
#get diagnosticity scores
toa_ref_epith_mesen <- get_toa_ref(gene_symbols = epith_mesen_ref_raw[,1],
                                   exp_treatment = epith_mesen_ref_raw[,2:11],
                                   exp_control = epith_mesen_ref_raw[,12:21])
}
