% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumanCD14CD16NegVsPos_8Reps_Wong}
\alias{HumanCD14CD16NegVsPos_8Reps_Wong}
\title{HumanCD14CD16NegVsPos_8Reps_Wong dataset}
\format{
An object of class \code{data.frame} with 3510 rows and 17 columns.
}
\usage{
HumanCD14CD16NegVsPos_8Reps_Wong
}
\description{
HumanCD14CD16NegVsPos_8Reps_Wong dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{internal}
