% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanCD14CD16NegVsPosVsDC_3Reps_Ziegler.R
\docType{data}
\name{HumanCD14CD16NegVsPosVsDC_3Reps_Ziegler}
\alias{HumanCD14CD16NegVsPosVsDC_3Reps_Ziegler}
\title{HumanCD14CD16NegVsPosVsDC_3Reps_Ziegler dataset}
\description{
HumanCD14CD16NegVsPosVsDC_3Reps_Ziegler dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{data}
