% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumanM1M2_3Reps_Martinez}
\alias{HumanM1M2_3Reps_Martinez}
\title{HumanM1M2_3Reps_Martinez dataset}
\format{
An object of class \code{data.frame} with 22283 rows and 7 columns.
}
\usage{
HumanM1M2_3Reps_Martinez
}
\description{
HumanM1M2_3Reps_Martinez dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{internal}
