% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanTransfacTELiS2019.R
\docType{data}
\name{HumanTransfacTELiS2019}
\alias{HumanTransfacTELiS2019}
\title{HumanTransfacTELiS2019 database}
\description{
HumanTransfacTELiS2019 database
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{data}
