% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_symbols.R
\name{set_symbols}
\alias{set_symbols}
\title{reformat gene symbols}
\usage{
set_symbols(symbols)
}
\arguments{
\item{symbols}{a character vector of gene symbols.}
}
\value{
a reformatted character vector of gene symbols.
}
\description{
reformats gene symbols such that space characters are omited and all remaining characters are upper case.
}
\examples{
#load example data
data("epith_mesen_ref_raw")
#clean symbols
epith_mesen_ref_raw$gene <- set_symbols(epith_mesen_ref_raw$gene)
}
