% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TAU_Trials3_Gene_CPM_Log2}
\alias{TAU_Trials3_Gene_CPM_Log2}
\title{TAU_Trials3_Gene_CPM_Log2 dataset}
\format{
An object of class \code{data.frame} with 60677 rows and 59 columns.
}
\usage{
TAU_Trials3_Gene_CPM_Log2
}
\description{
TAU_Trials3_Gene_CPM_Log2 dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{internal}
