% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TAUTrials2022BC_Intervention_Rm1BadQC_RmB14}
\alias{TAUTrials2022BC_Intervention_Rm1BadQC_RmB14}
\title{TAUTrials2022BC_Intervention_Rm1BadQC_RmB14 dataset}
\format{
An object of class \code{data.frame} with 26 rows and 2 columns.
}
\usage{
TAUTrials2022BC_Intervention_Rm1BadQC_RmB14
}
\description{
TAUTrials2022BC_Intervention_Rm1BadQC_RmB14 dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{internal}
