% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbm.R
\name{tfbm}
\alias{tfbm}
\title{Examine relative frequency of Transcription Factor Binding Motifs}
\usage{
tfbm(toa_boot_result, tfbm_ref)
}
\arguments{
\item{toa_boot_result}{a toa_boot result object produced using \code{toa_boot()}.}

\item{tfbm_ref}{a list object containing several tfbm reference data frames;
the first column of each data frame must contain gene symbols and be named "gene"}
}
\value{
tfbm returns an object of class "tfbm"

An object of class "tfbm" is a list containing:
\enumerate{
\item a results data frame.
\item a raw ratios data frame.
\item a list of input arguments.
}
}
\description{
performs a variant of transcript origin analysis aimed at determining whether the
frequency of Transcription Factor Binding Motif is significantly increased/decreased
among differentially expressed genes.
}
\examples{
\dontrun{
#' #load example data
data("TAU_Trials3_Gene_CPM_Log2")
data("TAUTrials2022BC_Intervention_Rm1BadQC_RmB14")
data("HumanM1M2_3Reps_Martinez")

#get DEG
DEG_result <- get_DEG(expression_data = TAU_Trials3_Gene_CPM_Log2,
exp_symbol_col = 1,
regressor_matrix = TAUTrials2022BC_Intervention_Rm1BadQC_RmB14,
reg_id_col = 1,
foldThreshDEG = 2,
screenSD = 0,
verbose = TRUE)

#get toa
toa_result <- toa(DEG_result = DEG_result,
ref = HumanM1M2_3Reps_Martinez,
type_1_cols = 2:4,
type_2_cols = 5:7)

#get bootstrapped stats
toa_boot_result <- toa_boot(toa_result)

#load a tfbm database from another repo (because it is >27MB)
library(Rfssa)
load_github_data("https://github.com/amanigaultw/TELiS/blob/main/HumanTransfacTELiS2019.RData")

#tfbm
tfbm_result <- tfbm(toa_boot_result, HumanTransfacTELiS2019)
View(tfbm_result$df_results)
}
}
