% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toa.R
\name{toa}
\alias{toa}
\title{Transcript Origin Analysis}
\usage{
toa(x, genes, toa_ref, cov = NULL, foldThreshDEG = 1.5)
}
\arguments{
\item{x}{A numeric vector of values to be used as the predictor variable of interest.}

\item{genes}{A numeric matrix if gene expression values, where columns are individual genes.}

\item{toa_ref}{A data frame containing gene symbols and diagnosticity scores; should be generated by using `get_toa_ref`}

\item{cov}{A numeric vector/matrix of values, where columns are individual covariate measures; if NULL, unadjusted regression coefficients (dif) are produced}

\item{foldThreshDEG}{the X-fold expression threshold to be exceeded for a given gene to be considered differentially expressed.}
}
\value{
A list object containing 2 data frames, including a summary table (df.means) and a dataframe containing individual gene symbol, regression coefficients, DEG, and toa_ref matching
}
\description{
`toa` performs transcript origin analysis
}
\examples{
#load example data
data("Chang")
data("toa_ref_epith_mesen")
toa_result <- toa(x = Chang$stress, genes = subset(Chang, select = -stress), toa_ref = toa_ref, foldThreshDEG = 1.25)
}
