% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toa_result_format.R
\name{toa_result_format}
\alias{toa_result_format}
\title{Format TOA results}
\usage{
toa_result_format(toa_boot_result)
}
\arguments{
\item{toa_boot_result}{a toa_boot result object produced using \code{toa_boot()}.#' @return toa returns an object of class "toa"}
}
\value{
a formated results matrix.
}
\description{
formats results from the bootstrapped transcript origin analysis
function to ease comparison with past/other work.
}
\examples{
\dontrun{
#load example data
data("TAU_Trials3_Gene_CPM_Log2")
data("TAUTrials2022BC_Intervention_Rm1BadQC_RmB14")
data("HumanM1M2_3Reps_Martinez")

#get DEG
DEG_result <- get_DEG(expression_data = TAU_Trials3_Gene_CPM_Log2,
                      regressor_matrix = TAUTrials2022BC_Intervention_Rm1BadQC_RmB14,
                      foldThreshDEG = 2)

#get toa
toa_result <- toa(DEG_result = DEG_result,
                  ref = HumanM1M2_3Reps_Martinez,
                  type_1_cols = 2:4,
                  type_2_cols = 5:7)

#get bootstrapped stats
toa_boot_result <- toa_boot(toa_result)

#generate a formated toa results table
pretty_results <- toa_result_format(toa_boot_result)
View(pretty_results)
}
}
