% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_roxygen_documentation.R
\docType{data}
\name{epith_mesen_ref_raw}
\alias{epith_mesen_ref_raw}
\title{epith_mesen_ref_raw dataset}
\description{
epith_mesen_ref_raw dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{data}
