% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DEG.R
\name{get_DEG}
\alias{get_DEG}
\title{get Differentially Expressed Genes}
\usage{
get_DEG(
  expression_data,
  exp_symbol_col = 1,
  regressor_matrix,
  reg_id_col = 1,
  foldThreshDEG = 1.5,
  screenSD = 0,
  verbose = FALSE
)
}
\arguments{
\item{foldThreshDEG}{the X-fold expression threshold to be exceeded for a given gene to be
considered differentially expressed.}

\item{screenSD}{minimum gene expression SD; genes that show expression variability below
threshold are excluded.}
}
\value{
returns an object of class "DEG"

An object of class "DEG" is a list containing:
\enumerate{
\item a data frame containing gene symbols, raw regression coefficients, and a DEG variable;
DEG values indicate whether a given gene showed expression values beyond threshold (1 = increased; -1 = decreased; 0 = below threshold).
\item a dataframe generated by transposing the expression_data and merging with it with the regressor_matrix.
\item a list of input arguments.
}
}
\description{
screens the \code{expression_data}, then merges it with the \code{regressor_matrix} and performs regressions to determine which gene show differential expression beyond \code{foldThreshDEG}.
can adjust for covariates using \code{cov}.
}
\examples{
\dontrun{
#load example data
data("TAU_Trials3_Gene_CPM_Log2")
data("TAUTrials2022BC_Intervention_Rm1BadQC_RmB14")

#get DEG
DEG_result <- get_DEG(expression_data = TAU_Trials3_Gene_CPM_Log2,
exp_symbol_col = 1,
regressor_matrix = TAUTrials2022BC_Intervention_Rm1BadQC_RmB14,
reg_id_col = 1,
foldThreshDEG = 2,
screenSD = 0,
verbose = TRUE)

#view results
table(DEG_result$df_DEG$DEG)
}
}
