% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DEG.R
\name{get_DEG}
\alias{get_DEG}
\title{get Differentially Expressed Genes}
\usage{
get_DEG(x, genes, cov = NULL, foldThreshDEG = 1.5)
}
\arguments{
\item{x}{a numeric vector of values; the predictor variable of interest.}

\item{genes}{a numeric matrix of gene expression values, where columns are individual genes.}

\item{cov}{a numeric vector/matrix of values, where columns are individual covariate measures; if NULL, unadjusted regression coefficients (dif) are produced.}

\item{foldThreshDEG}{the X-fold expression threshold to be exceeded for a given gene to be considered differentially expressed.}
}
\value{
a data frame containing gene symbols, raw regression coefficients, and a DEG variable; DEG values indicate whether a given gene showed expression values beyond threshold (1 = increased; -1 = decreased; 0 = below threshold).
}
\description{
\code{get_DEG} performs regressions to determine which gene show differential expression as a function of x; can adjust for covariates using \code{cov}.
}
\examples{
#load example data
data("Chang")
data("epith_mesen_ref_raw")
#get_toa_ref
toa_ref_epith_mesen <- get_toa_ref(gene_symbols = epith_mesen_ref_raw[,1], exp_treatment = epith_mesen_ref_raw[,2:11], exp_control = epith_mesen_ref_raw[,12:21])
}
