% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumanCD14CD16NegVsPosVsDC_3Reps_Ziegler}
\alias{HumanCD14CD16NegVsPosVsDC_3Reps_Ziegler}
\title{HumanCD14CD16NegVsPosVsDC_3Reps_Ziegler dataset}
\format{
An object of class \code{data.frame} with 21026 rows and 7 columns.
}
\usage{
HumanCD14CD16NegVsPosVsDC_3Reps_Ziegler
}
\description{
HumanCD14CD16NegVsPosVsDC_3Reps_Ziegler dataset
}
\author{
Andrew Manigault \email{andrew_manigault@brown.edu}
}
\keyword{internal}
