% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbm.R
\name{tfbm}
\alias{tfbm}
\title{Transcript Origin Analysis for Transcription Factor Binding Motifs}
\usage{
tfbm(x, genes, tfbm_ref, cov = NULL, foldThreshDEG = 1.5, lite = FALSE)
}
\arguments{
\item{x}{a numeric vector of values; the predictor variable of interest.}

\item{genes}{a numeric matrix of gene expression values, where columns are individual genes.}

\item{tfbm_ref}{a list object containing several tfbm reference data frames;
the first column of each data frame must contain gene symbols and be named "gene"}

\item{cov}{a numeric vector/matrix of values, where columns are individual covariate measures;
if NULL, unadjusted regression coefficients (dif) are produced.}

\item{foldThreshDEG}{the X-fold expression threshold to be exceeded for a given gene to be
considered differentially expressed.}

\item{lite}{a bool indicating whether a stripped down version of the function
should be run; FALSE by default.}
}
\value{
tfbm returns an object of class "tfbm"

An object of class "tfbm" is a list containing:
\enumerate{
\item a results data frame.
\item a raw coefficients data frame.
\item a raw ratios data frame.
\item a list of input arguments.
}
}
\description{
performs transcript origin analysis for Transcription Factor Binding Motif
without bootstrapped estimates of tfbm ratios.
}
\examples{
\dontrun{
#load example data
data("Chang")

#load a tfbm database from another repo (because it is >27MB)
library(Rfssa)
load_github_data("https://github.com/amanigaultw/TELiS/blob/main/HumanTransfacTELiS2019.RData")

#tfbm
tfbm_result <- tfbm(x <- Chang[,1],
                    genes <- Chang[,-1],
                    tfbm_ref = HumanTransfacTELiS2019,
                    cov = NULL,
                    foldThreshDEG = 1.25)
}
}
