% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbm_boot.R
\name{tfbm_boot}
\alias{tfbm_boot}
\title{bootstrapped Transcript Origin Analysis for Transcription Factor Binding Motifs}
\usage{
tfbm_boot(tfbm, n_boot = 200, progress = TRUE)
}
\arguments{
\item{tfbm}{a tfbm result object produced using \code{tfbm()}.}

\item{n_boot}{the number of bootstrap sample to run.}

\item{progress}{a bool indicating whether a progress bar should be shown.}
}
\value{
a list object containing:
\enumerate{
\item a results data frame.
\item a matrix containing bootstrapped mean diagnosticity score estimates \code{boot}.
\item a list of input arguments.
}
}
\description{
performs transcript origin analysis for Transcription Factor Binding Motif, including bootstrapped estimation of tfbm ratios; will use n-1 available CPU cores by default.
}
\examples{
#load example data
data("Chang")

#load a tfbm database from another repo (because it is >27MB)
#library(Rfssa)
#load_github_data("https://github.com/amanigaultw/TELiS/blob/main/HumanTransfacTELiS2019.RData")

#tfbm
#tfbm_result <- tfbm(x <- Chang[,1],
#                    genes <- Chang[,-1],
#                    tfbm_ref = HumanTransfacTELiS2019,
#                    cov = NULL,
#                    foldThreshDEG = 1.25)
#tfbm_boot
#tfbm_boot_results <- tfbm_boot(tfbm_result)
}
