% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toa.R
\name{toa}
\alias{toa}
\title{Transcript Origin Analysis}
\usage{
toa(x, genes, toa_ref, cov = NULL, foldThreshDEG = 1.5)
}
\arguments{
\item{x}{a numeric vector of values; the predictor variable of interest.}

\item{genes}{a numeric matrix of gene expression values, where columns are individual genes.}

\item{toa_ref}{a data frame containing gene symbols and diagnosticity scores; should be generated by using \code{get_toa_ref()}.}

\item{cov}{a numeric vector/matrix of values, where columns are individual covariate measures; if NULL, unadjusted regression coefficients (dif) are produced.}

\item{foldThreshDEG}{the X-fold expression threshold to be exceeded for a given gene to be considered differentially expressed.}
}
\value{
toa returns an object of class "toa"

An object of class "toa" is a list containing:
\enumerate{
\item a results data frame.
\item a raw coefficients data frame.
\item a list of input arguments.
}
}
\description{
performs transcript origin analysis without bootstrapped estimates of mean diagnosticity scores.
}
\examples{
#load example data
data("Chang")
data("epith_mesen_ref_raw")
#get diagnosticity scores
toa_ref_epith_mesen <- get_toa_ref(gene_symbols = epith_mesen_ref_raw[,1],
                                   exp_treatment = epith_mesen_ref_raw[,2:11],
                                   exp_control = epith_mesen_ref_raw[,12:21])
#toa
toa_result <- toa(x = Chang$stress,
                  genes = subset(Chang, select = -stress),
                  toa_ref = toa_ref_epith_mesen,
                  foldThreshDEG = 1.25)
}
