% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toa.R
\name{toa}
\alias{toa}
\title{Transcript Origin Analysis}
\usage{
toa(x, genes, toa_ref, cov = NULL, foldThreshDEG = 1.5)
}
\arguments{
\item{x}{a numeric vector of values; the predictor variable of interest.}

\item{genes}{a numeric matrix of gene expression values, where columns are individual genes.}

\item{toa_ref}{a data frame containing gene symbols and diagnosticity scores; should be generated by using \code{\link{get_toa_ref}}.}

\item{cov}{a numeric vector/matrix of values, where columns are individual covariate measures; if NULL, unadjusted regression coefficients (dif) are produced.}

\item{foldThreshDEG}{the X-fold expression threshold to be exceeded for a given gene to be considered differentially expressed.}
}
\value{
a list object containing:
\enumerate{
  \item a summary table \code{df_results}.
  \item a raw coefficients data frame \code{df_DEG}.
}
}
\description{
\code{toa} performs transcript origin analysis without bootstrapped estimates of mean diagnosticity scores.
}
\examples{
#load example data
data("Chang")
data("toa_ref_epith_mesen")
#toa
toa_result <- toa(x = Chang$stress,
                  genes = subset(Chang, select = -stress),
                  toa_ref = toa_ref_epith_mesen,
                  foldThreshDEG = 1.25)
}
