% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toa_boot.R
\name{toa_boot}
\alias{toa_boot}
\title{bootstrapped Transcript Origin Analysis}
\usage{
toa_boot(
  x,
  genes,
  toa_ref,
  cov = NULL,
  foldThreshDEG = 1.5,
  n_boot = 200,
  show_progress = TRUE
)
}
\arguments{
\item{x}{a numeric vector of values; the predictor variable of interest.}

\item{genes}{a numeric matrix of gene expression values, where columns are individual genes.}

\item{toa_ref}{a data frame containing gene symbols and diagnosticity scores; should be generated by using \code{get_toa_ref}.}

\item{cov}{a numeric vector/matrix of values, where columns are individual covariate measures; if NULL, unadjusted regression coefficients (dif) are produced.}

\item{foldThreshDEG}{the X-fold expression threshold to be exceeded for a given gene to be considered differentially expressed.}

\item{n_boot}{the number of bootstrap sample to run}

\item{show_progress}{a bool indicating whether a progress bar should be shown}
}
\value{
a list object containing:
\enumerate{
  \item a summary table \code{df_results}.
  \item a matrix containing bootstrapped mean diagnosticity score estimates \code{boot}.
}
}
\description{
\code{toa_boot} performs transcript origin analysis, including bootstrapped estimates of mean diagnosticity scores; will use n-1 available CPU cores by default.
}
\examples{
#load example data
data("Chang")
data("toa_ref_epith_mesen")
#toa_boot
#toa_boot_result <- toa_boot(x = Chang$stress,
                             genes = subset(Chang, select = -stress),
                             toa_ref = toa_ref_epith_mesen,
                             foldThreshDEG = 1.25)
}
