% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toa_boot.R
\name{toa_boot}
\alias{toa_boot}
\title{bootstrapped Transcript Origin Analysis}
\usage{
toa_boot(toa, n_boot = 200, progress = TRUE)
}
\arguments{
\item{toa}{a toa result object produced using \code{toa()}.}

\item{n_boot}{the number of bootstrap sample to run.}

\item{progress}{a bool indicating whether a progress bar should be shown.}
}
\value{
a list object containing:
\enumerate{
  \item a results data frame.
  \item a matrix containing bootstrapped mean diagnosticity score estimates \code{boot}.
}
}
\description{
performs transcript origin analysis, including bootstrapped estimation of mean diagnosticity scores; will use n-1 available CPU cores by default.
}
\examples{
#load example data
data("Chang")
data("epith_mesen_ref_raw")
#get diagnosticity scores
toa_ref_epith_mesen <- get_toa_ref(gene_symbols = epith_mesen_ref_raw[,1],
                                   exp_treatment = epith_mesen_ref_raw[,2:11],
                                   exp_control = epith_mesen_ref_raw[,12:21])
#toa
toa_result <- toa(x = Chang$stress,
                  genes = subset(Chang, select = -stress),
                  toa_ref = toa_ref_epith_mesen,
                  foldThreshDEG = 1.25)
#toa_boot
#toa_boot_result <- toa_boot(toa = toa_result,
#                            n_boot = 400) #few bootstrapped samples for illustration
}
